Fix commands now that Discord is not down and I can test it

+++ MrClean/Commands/Filters/CreateMessageFilterCommand.cs
@@ -52,7 +52,7 @@ public class CreateMessageFilterCommand : ISlashCommandProvider
         await command.DeferAsync();
         
         var pattern = command.GetOption<string>("pattern");
-        var delay = Math.Clamp(command.GetOption<int>("delay"), 0, 120);
+        var delay = (int) Math.Clamp(command.GetOption<long>("delay"), 0, 120);
         var channel = command.GetOption<SocketGuildChannel>("repost-channel");
 
         await using var context = await _factory.CreateDbContextAsync();
